Move tiles tests to own file
