+++ tests/test_api.py
@@ -49,16 +49,11 @@ from pygeoapi.api.coverages import get_collection_coverage
 from pygeoapi.api.itemtypes import (
     get_collection_schema, get_collection_queryables, get_collection_item,
     get_collection_items, manage_collection_item)
-from pygeoapi.api.tiles import (
-    get_collection_tiles, tilematrixset, tilematrixsets,
-)
 from pygeoapi.util import (yaml_load, get_crs_from_uri,
                            get_api_rules, get_base_url)
 
 from .util import (get_test_file_path, mock_request,
                    mock_flask, mock_starlette, mock_api_request)
-
-from pygeoapi.models.provider.base import TileMatrixSetEnum
 
 LOGGER = logging.getLogger(__name__)
 
@@ -621,53 +616,6 @@ def test_conformance(config, api_):
 
     req = mock_request({'f': 'html'})
     rsp_headers, code, response = api_.conformance(req)
-    assert rsp_headers['Content-Type'] == FORMAT_TYPES[F_HTML]
-    # No language requested: should be set to default from YAML
-    assert rsp_headers['Content-Language'] == 'en-US'
-
-
-def test_tilematrixsets(config, api_):
-    req = mock_api_request()
-    rsp_headers, code, response = tilematrixsets(api_, req)
-    root = json.loads(response)
-
-    assert isinstance(root, dict)
-    assert 'tileMatrixSets' in root
-    assert len(root['tileMatrixSets']) == 2
-    assert 'http://www.opengis.net/def/tilematrixset/OGC/1.0/WorldCRS84Quad' \
-           in root['tileMatrixSets'][0]['uri']
-    assert 'http://www.opengis.net/def/tilematrixset/OGC/1.0/WebMercatorQuad' \
-           in root['tileMatrixSets'][1]['uri']
-
-    req = mock_api_request({'f': 'html'})
-    rsp_headers, code, response = tilematrixsets(api_, req)
-    assert rsp_headers['Content-Type'] == FORMAT_TYPES[F_HTML]
-    # No language requested: should be set to default from YAML
-    assert rsp_headers['Content-Language'] == 'en-US'
-
-
-def test_tilematrixset(config, api_):
-    req = mock_api_request()
-
-    enums = [e.value for e in TileMatrixSetEnum]
-    enum = None
-
-    for e in enums:
-        enum = e.tileMatrixSet
-        rsp_headers, code, response = tilematrixset(api_, req, enum)
-        root = json.loads(response)
-
-        assert isinstance(root, dict)
-        assert 'id' in root
-        assert root['id'] == enum
-        assert 'tileMatrices' in root
-        assert len(root['tileMatrices']) == 30
-
-    rsp_headers, code, response = tilematrixset(api_, req, 'foo')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    req = mock_api_request({'f': 'html'})
-    rsp_headers, code, response = tilematrixset(api_, req, enum)
     assert rsp_headers['Content-Type'] == FORMAT_TYPES[F_HTML]
     # No language requested: should be set to default from YAML
     assert rsp_headers['Content-Language'] == 'en-US'
@@ -1557,25 +1505,6 @@ def test_get_collection_coverage(config, api_):
     # rsp_headers, code, response = get_collection_coverage(api_ req, 'cmip5')
     #
     # assert code == HTTPStatus.NO_CONTENT
-
-
-def test_get_collection_tiles(config, api_):
-    req = mock_api_request()
-    rsp_headers, code, response = get_collection_tiles(api_, req, 'obs')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    rsp_headers, code, response = get_collection_tiles(
-        api_, req, 'naturalearth/lakes')
-    assert code == HTTPStatus.OK
-
-    # Language settings should be ignored (return system default)
-    req = mock_api_request({'lang': 'fr'})
-    rsp_headers, code, response = get_collection_tiles(
-        api_, req, 'naturalearth/lakes')
-    assert rsp_headers['Content-Language'] == 'en-US'
-    content = json.loads(response)
-    assert len(content['links']) > 0
-    assert len(content['tilesets']) > 0
 
 
 def test_validate_bbox():
